Отрефакторил сетевые сервисы
