+++ FakeNFT/Epics/Catalog/CatalogScreen/NetworkService/CatalogNetworkService.swift
@@ -50,21 +50,3 @@ final class CatalogNetworkService {
         }
     }
 }
-
-final class CatalogRequest: NetworkRequest {
-    var httpMethod: HttpMethod
-    var endpoint: URL? {
-        var urlComponents = URLComponents(string: "https://\(CatalogConstants.host.rawValue)/api/v1/collections")
-        urlComponents?.queryItems = [
-            URLQueryItem(name: "page", value: nextPage),
-            URLQueryItem(name: "limit", value: "10")
-        ]
-        return urlComponents?.url
-    }
-    private var nextPage: String
-    
-    init(nextPage: String, httpMethod: HttpMethod) {
-        self.nextPage = nextPage
-        self.httpMethod = httpMethod
-    }
-}
