+++ FakeNFT/Epics/Catalog/CollectionScreen/NetworkService/AuthorNetworkService/AuthorModel.swift
