+++ FakeNFT/Epics/Catalog/CollectionScreen/NetworkService/NftNetworkService/NftModel.swift
